+++ plugins/source/azure/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this provider will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.10](https://github.com/cloudquery/cloudquery/compare/plugins-source-azure-v1.4.9...plugins-source-azure-v1.4.10) (2022-11-30)
+
+
+### Bug Fixes
+
+* **deps:** Update module github.com/cloudquery/plugin-sdk to v1.10.0 ([#5153](https://github.com/cloudquery/cloudquery/issues/5153)) ([ea1f77e](https://github.com/cloudquery/cloudquery/commit/ea1f77e910f430287600e74cedd7d3f4ae79eb18))
+* **deps:** Update plugin-sdk for azure to v1.8.1 ([#5033](https://github.com/cloudquery/cloudquery/issues/5033)) ([3c40cb5](https://github.com/cloudquery/cloudquery/commit/3c40cb5942692f7d808834950409cee89fe49fd7))
+* **deps:** Update plugin-sdk for azure to v1.8.2 ([#5075](https://github.com/cloudquery/cloudquery/issues/5075)) ([029b138](https://github.com/cloudquery/cloudquery/commit/029b13867c2d0cd1e119e6a8a1ed4fa8c30c6bdc))
+* **deps:** Update plugin-sdk for azure to v1.9.0 ([#5093](https://github.com/cloudquery/cloudquery/issues/5093)) ([ff9427f](https://github.com/cloudquery/cloudquery/commit/ff9427f8514615d611e09a2f3a4e2bf1b46170fc))
+
 ## [1.4.9](https://github.com/cloudquery/cloudquery/compare/plugins-source-azure-v1.4.8...plugins-source-azure-v1.4.9) (2022-11-23)